openclaw/parser/requirements.txt
# Doc Parser Dependencies
pymupdf>=1.23.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pytesseract>=0.3.10
pillow>=10.0.0
uvicorn>=0.24.0
fastapi>=0.104.0
